fix: pass --repo flag to preview and action commands
Pass the --repo flag to the preview and gh commands called via
keybindings. This only caused issues when the CWD was in a different
repository than the one specified via the --repo flag. In those cases,
the preview and actions would be for the current repo (based on CWD),
rather than the one specified by the flag.