fix: pass repo flag to preview and keybinding commands (#8)
Resolve issue when the current working directory (CWD) was in a
different repository than the one specified via the `--repo` flag. In
those cases, the commands for the preview and keybinding actions would
be called on the current repo (based on CWD), rather than the one
specified by the flag.