fix: blocked process when opening item in browser
Use a background process when executing commands that open an item in
the browser. For example:

        $ gh pr view --web <number>

When a new browser process was created, `gh=fzf` became unresponsive
until the browser was terminated. The `gh-fzf` process was not blocked
if the browser was already open, because it would add the item as a new
tab instead of creating a new process.