feat(issue): add options for naming linked branch
Add environment variables that prepend info to branch names when they
are created and linked to the selected issue. Spaces in the variables
are removed.

- GH_FZF_BRANCH_PREFIX: a string at the beginning of the branch name
- GH_FZF_BRANCH_ADD_ISSUE_NUMBER: When non-empty, the issue number is
  added after the prefix (if specified), and the value of this
  environment variable is added after the number.

For example, the branch name will be "benelan/123-fix-bug" if you enter
"fix bug" when prompted and have the following variables set:

    export GH_FZF_BRANCH_PREFIX="$(
        git config --global github.user || echo "benelan"
     )/"
    export GH_FZF_BRANCH_ADD_ISSUE_NUMBER="-"

Note that spaces in the prompt's value are replaced with hyphens.
Another example:

    export GH_FZF_BRANCH_PREFIX="issue-"
    export GH_FZF_BRANCH_ADD_ISSUE_NUMBER=" "

This will result in the branch name "issue-123" if you press enter
without adding a value to the prompt, because spaces in the environment
variables are removed.