fix: config issues when user hasn't set FZF_DEFAULT_OPTS
The important fzf options that I ensured would not be overridden by a
user's FZF_DEFAULT_OPTS were not being applied if the environment
variable was unset. The important options are:

    --no-select-1 --ansi --no-multi

This caused ansi escape sequences to not be handled properly, leading to
garbled text. Interestingly, setting the env var to any value would fix
the issue. For example:

    $ FZF_DEFAULT_OPTS="" gh fzf issue

Shells are weird.