feat(run): notify desktop when completed (experimental)
Add a util subcommand for watching a GitHub Action run and displaying a
desktop notification on completion. Usage:

    $ gh fzf util notify-run-completed <run-id>

The supported notification programs (attempted in order) are:

- dunstify
- notify-send
- osascript (untested so far)

The util doesn't have a builtin keybinding for now, but can be enabled
in a gh config alias:

    # ~/.config/gh/config.yml
    aliases:
        runs: |
            !FZF_DEFAULT_OPTS="$FZF_DEFAULT_OPTS
                --bind='alt-n:execute-silent(
                    gh fzf util notify-run-completed {-1} &
                )'
            " gh fzf run $*

An icon can be displayed next to the notification if you are using
dunstify or notify-send. Set the `GH_FZF_NOTIFY_ICON` environment
variable to specify the icon path:

    export GH_FZF_NOTIFY_ICON="$XDG_DATA_HOME/icons/Gruvbox/apps/scalable/github.svg"

WARNING: Subcommands of `gh fzf util` are considered "internal", meaning
their issues may be closed with: Whatever, I do what I want ¯\_(ツ)_/¯