refactor(milestone): add option for cmd to open url
Check if `GH_FZF_OPEN_CMD` is set, and if so, use the environment
variable's value to open a milestone's URL in the browser. Defaults to
the `BROWSER` environment variable if set, otherwise it uses the
operating system's open command. Any future commands that fetch the list
with `gh api` will use this environment variable too.